git-svn-id: svn+ssh://scheduler.csc.calpoly.edu/repo@994 13b268f7-8392-4c9d-bc7d-1e80a86bc2f9

+++ implementation/model/java/edu/calpoly/csc/scheduler/model/db/Time.java
@@ -14,7 +14,7 @@ import java.io.Serializable;
 public class Time implements Comparable<Time>, Serializable
 {
    /** Serial Version UID */
-   public static final int serialVersionUID = 42;
+   public static final long serialVersionUID = 42;
 
    /**
     * A list of all the half-hour time slots in a day.
